+++ apps/dashboard/src/polyfills.tsx
@@ -1,3 +1,4 @@
+import * as util from "util";
 import { registerErrorSink } from "@stackframe/stack-shared/dist/utils/errors";
 import * as Sentry from "@sentry/nextjs";
 
@@ -7,6 +8,10 @@ const sentryErrorSink = (location: string, error: unknown) => {
 
 export function ensurePolyfilled() {
   registerErrorSink(sentryErrorSink);
+  // not all environments have default options for util.inspect
+  if ("inspect" in util && "defaultOptions" in util.inspect) {
+    util.inspect.defaultOptions.depth = 8;
+  }
 }
 
 ensurePolyfilled();