+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stackframe/stack-backend
 
+## 2.5.16
+
+### Patch Changes
+
+- @stackframe/stack-emails@2.5.16
+- @stackframe/stack-shared@2.5.16
+
 ## 2.5.15
 
 ### Patch Changes