+++ docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @stackframe/docs
 
+## 2.5.16
+
 ## 2.5.15
 
 ### Patch Changes