+++ examples/demo/CHANGELOG.md
@@ -1,5 +1,13 @@
 # demo-app
 
+## 2.5.16
+
+### Patch Changes
+
+- @stackframe/stack@2.5.16
+- @stackframe/stack-shared@2.5.16
+- @stackframe/stack-ui@2.5.16
+
 ## 2.5.15
 
 ### Patch Changes